backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.4
langchain-core==0.1.10
langchain-community==0.0.16
langchain-openai==0.0.5
chromadb==0.4.22
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib==1.7.4
aiosqlite==0.19.0
sqlalchemy==2.0.25
httpx==0.26.0
streamlit==1.30.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-docx==1.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
